ManagedRandomAccessFile/OutputStream to read/write to .Net IO.Streams
Adds ManagedRandomAccessFile and ManagedOutputStream, implementations of
the RandomAccessFile and OutputStream C++ interfaces that call back into
C# to read/write to a .Net IO.Stream.

This seems to be on par with performance from letting the C++ side
manage the file itself, that is pinvoke overhead appears minimal
relative to IO costs. The TestFloatTimeSeries benchmark has been updated
to write out by passing a filepath to C++, and by using
ManagedOutputStream.

Example output of TestFloatTimeSeries (with a slightly larger data size
than default):

 Generating data...
 Generated 31,200,000 rows in 0.78 sec

 Saving to CSV
 Saved to CSV (1,262,808,034 bytes) in 43.87 sec

 Saving to CSV.GZ
 Saved to CSV (317,658,548 bytes) in 92.55 sec

 Saving to Parquet
 Saved to Parquet (168,610,125 bytes) in 1.06 sec

 Saving to Parquet.Chunked (by date)
 Saved to Parquet.Chunked (361,316,340 bytes) in 6.74 sec

 Saving to Parquet.RowOriented
 Saved to Parquet.RowOriented (168,510,739 bytes) in 1.98 sec

 Saving to Parquet.FileStream
 Saved to Parquet.FileStream (168,609,640 bytes) in 1.05 sec

 Saving to Parquet.Chunked.FileStream (by date)
 Saved to Parquet.Chunked.FileStream (361,316,332 bytes) in 5.14 sec

 Saving to Parquet.RowOriented.FileStream
 Saved to Parquet.RowOriented.FileStream (168,510,254 bytes) in 2.02 sec

 Saving to Parquet.NET
 Saved to Parquet.NET (261,420,477 bytes) in 11.60 sec

As can be seen the FileStream tests are on par with using C++ directly,
and are all much faster than Parquet.NET.