ParquetFileReader::Open takes a RandomAccessFile not an InputStream
Noticed this typing inconsistency when looking at doing a bigger change
(which I'll put in a separate PR). In cpp ParquertFileRead::Open takes a
RandomAccessFile, and a BufferReader inherits from RandomAccessFile. But
the C# code was saying that Open took an InputStream (that's less
specific than RandomAccessFile), and that BufferReader was an
InputStream (while it inherits directly from RandomAccessFile, which
itself has a multiple inheritance on InputStream).