Do not run CI for PRs where both branches are from the same repo (#143)
When a pull request is created by comparing branches that are on the same repo, duplicate CI runs could be created: one for the branch push, and one for the pull request. PR #100 avoided that by running the CI for all PRs, and for pushes to the master branch only.

This PR instead lets the CI run for any branch push, and only for PRs originating from a fork. This will reinstate the ability to run the CI for each branch push to a fork, whilst preserving it for each PR originating from a fork.