Fix CI with .NET >= 5.0.200
NuGet sources are specified in the right order in the top-level config
and the implicit restore done during the package add will use it instead
of a unique source specified on the command line.

Before .NET 5.0.200, the implicit restore was also considering the
nuget.org source, so this problem never occured.