Allow specifying column definitions for a ParquetRowWriter (#284)
* Allow providing an array of Column instances to CreateRowWriter

This gives users more control over the logical type mappings for columns,
eg. to specify IsAdjustedToUtc for Timestamps, and works when using a tuple
type where we can't add attributes to fields.

* Move column mapping out of write delegate cache

This prevents getting the incorrect columns when re-using the write
delegate and columns have been manually specified.

Introduces a new MappedField struct to replace the previous tuple to
make this more manageable.

* Comment fixes