Avoid use of ManagedOutputStream/RandomAccessFile after GC
This fixes crashes when users don't keep a reference to a
ManagedOutputStream or ManagedRandomAccessFile while they are in use.
They should still be used with a using statement so that the handle
is freed immediately when no longer needed though.

A reference to the managed objects is captured in the handle free functions
so that they are kept alive until after the handle is freed.
ParquetFileReader and Writer now also keep references to the
RandomAccessFile and OutputStream to ensure they aren't GC'd until after
they are no longer needed.