Add prebuilt devcontainer (#369)
* Use vcpkg.json only

This is necessary to be able to build ParquetSharpNative offline with
a populated vcpkg cache.

* Ensure vcpkg builtin registry is up-to-date in CI runners

If we reference a baseline that is too recent, some CI runner images
will not have it. With this step we ensure it won't be the case.

This used to be done via vcpkg-configuration.json, but we couldn't rely
on it for offline devcontainer usage.

* Improve Powershell build script

Format via Powershell VS Code extension and fix
linting issues.

* Build native lib in both Debug and Release by default (unless in the CI)

* Add devcontainer

* Add devcontainer workflow

* Update documentation about building ParquetSharp

* Add solution to .gitignore